Added backend Rust book recommendation to pre-reqs
the suggested book introduces backend development, including distributed systems,  in Rust. I don't think it's entirely necessary and may be overkill, which is why I added "optional but highly recommended" but it will definitely be great prep for individuals who do not have experience in distributed systems, and so it will build strong contributors.